backend/requirements.txt
asgiref==3.7.2
certifi==2023.11.17
cfgv==3.4.0
charset-normalizer==3.3.2
click==8.1.7
distlib==0.3.7
Django==4.2.7
django-cors-headers==4.3.1
djangorestframework==3.14.0
docker==6.1.3
drf-yasg==1.21.7
filelock==3.13.1
flake8==6.1.0
fsspec==2023.10.0
fuzzywuzzy==0.18.0
gunicorn==21.2.0
huggingface-hub==0.19.4
identify==2.5.32
idna==3.6
inflection==0.5.1
isort==5.12.0
Jinja2==3.1.2
joblib==1.3.2
Levenshtein==0.23.0
lightgbm==4.1.0
MarkupSafe==2.1.3
mccabe==0.7.0
mpmath==1.3.0
mypy-extensions==1.0.0
networkx==3.2.1
nodeenv==1.8.0
numpy==1.26.2
packaging==23.2
pandas==1.4.4
pathspec==0.11.2
platformdirs==4.0.0
psycopg2-binary==2.9.9
pycodestyle==2.11.1
pyflakes==3.1.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-Levenshtein==0.23.0
pytz==2023.3.post1
PyYAML==6.0.1
rapidfuzz==3.5.2
regex==2023.10.3
requests==2.31.0
safetensors==0.4.1
scikit-learn==1.1.3
scipy==1.11.4
six==1.16.0
sqlparse==0.4.4
sympy==1.12
threadpoolctl==3.2.0
tokenizers==0.15.0
torch==2.0.0
tqdm==4.66.1
transformers==4.35.2
typing_extensions==4.8.0
uritemplate==4.1.1
urllib3==2.1.0
virtualenv==20.25.0
websocket-client==1.6.4
